chore(EDCO-34): file tree refactoring
