+++ src/components/Footer/Footer.tsx
@@ -1,112 +1,10 @@
-import { Box, Container, Link, Stack, Typography } from '@mui/material';
+import { Box, Container } from '@mui/material';
 import { FundedByNextGenerationEU } from '../../assets/FundedByNextGenerationEU';
-import { LogoPagoPACompany } from '../../assets/LogoPagoPACompany';
 import { type Generic } from '../../types/components';
-import { hrefNoOp, isJSX } from '../../utils/index';
-import { EIcon, type EIconProps } from '../EIcon';
-import {
-  type CompanyLinkType,
-  type PreLoginFooterLinksType,
-  type PreLoginFooterSingleSectionType,
-} from './index';
-import { LangSwitch, type LangSwitchProps } from './LangSwitch';
-
-export const LegalInfo = ({
-  data,
-}: {
-  data?: string | Generic | Generic[];
-}) => (
-  <Box
-    sx={{
-      borderTop: 1,
-      borderColor: 'divider',
-      backgroundColor: 'background.paper',
-    }}
-  >
-    <Container sx={{ px: 2, py: 2 }}>
-      <Typography
-        color="text.primary"
-        variant="caption"
-        textAlign="center"
-        component={isJSX(data) ? 'div' : 'p'}
-      >
-        {data}
-      </Typography>
-    </Container>
-  </Box>
-);
-
-export const FooterColumn = ({
-  data,
-  companyLink,
-  icons,
-}: {
-  data: PreLoginFooterSingleSectionType;
-  companyLink?: CompanyLinkType;
-  icons?: EIconProps[];
-}) => (
-  <Stack spacing={2} display="flex" alignItems={{ xs: 'center', sm: 'start' }}>
-    {data?.title && <Typography variant="overline">{data.title}</Typography>}
-
-    {companyLink && (
-      <Link
-        component="button"
-        role="link"
-        aria-label={companyLink.ariaLabel}
-        href={companyLink.href ?? hrefNoOp}
-        sx={{ display: 'inline-flex' }}
-      >
-        <LogoPagoPACompany />
-      </Link>
-    )}
-
-    {icons && (
-      <Stack
-        aria-label="i nostri social"
-        spacing={{ xs: 3, sm: 1.5, lg: 3 }}
-        direction="row"
-        component="ul"
-        alignItems={{ xs: 'center', sm: 'start' }}
-        sx={{ padding: 0, mt: 0.5, listStyle: 'none' }}
-      >
-        {icons?.map(
-          ({ icon, href = hrefNoOp, onClick, ...rest }, i) =>
-            icon && (
-              <li key={i}>
-                <Link href={href}>
-                  <EIcon icon={icon} color="text.primary" {...rest} />
-                </Link>
-              </li>
-            )
-        )}
-      </Stack>
-    )}
-
-    <Stack
-      component="ul"
-      alignItems={{ xs: 'center', sm: 'start' }}
-      textAlign={{ xs: 'center', sm: 'left' }}
-      sx={{ padding: 0, listStyle: 'none' }}
-    >
-      {data?.links.map(({ href = hrefNoOp, label, ariaLabel, onClick }, i) => (
-        <li key={i}>
-          <Link
-            aria-label={ariaLabel}
-            component="a"
-            href={href}
-            onClick={onClick}
-            underline="none"
-            color="text.primary"
-            sx={{ display: 'inline-block', py: 0.5 }}
-            variant="subtitle2"
-          >
-            {label}
-          </Link>
-        </li>
-      ))}
-    </Stack>
-  </Stack>
-);
+import { FooterColumn } from './components/FooterColumn';
+import { LangSwitch, type LangSwitchProps } from './components/LangSwitch';
+import { LegalInfo } from './components/LegalInfo';
+import { type CompanyLinkType, type PreLoginFooterLinksType } from './types';
 
 export interface FooterProps extends LangSwitchProps {
   companyLink: CompanyLinkType;
